+++ packages/backend/src/models/ProjectModel/ProjectModel.ts
@@ -21,7 +21,6 @@ import {
     ProjectMemberRole,
     ProjectSummary,
     ProjectType,
-    SemanticLayerType,
     SnowflakeAuthenticationType,
     SnowflakeCredentials,
     SpaceSummary,
@@ -41,10 +40,6 @@ import {
     isExploreError,
     sensitiveCredentialsFieldNames,
     sensitiveDbtCredentialsFieldNames,
-    type CubeSemanticLayerConnection,
-    type DbtSemanticLayerConnection,
-    type SemanticLayerConnection,
-    type SemanticLayerConnectionUpdate,
 } from '@lightdash/common';
 import {
     WarehouseCatalog,
@@ -392,19 +387,6 @@ export class ProjectModel {
                 throw new UnexpectedServerError('Could not save credentials.');
             }
 
-            let encryptedSemanticLayerConnection: Buffer | null;
-            try {
-                encryptedSemanticLayerConnection = data.semanticLayerConnection
-                    ? this.encryptionUtil.encrypt(
-                          JSON.stringify(data.semanticLayerConnection),
-                      )
-                    : null;
-            } catch (e) {
-                throw new UnexpectedServerError(
-                    'Could not encrypt semantic layer connection credentials.',
-                );
-            }
-
             // Make sure the project to copy exists and is owned by the same organization
             const copiedProjects = data.upstreamProjectUuid
                 ? await trx('projects')
@@ -423,7 +405,7 @@ export class ProjectModel {
                             ? copiedProjects[0].project_uuid
                             : null,
                     dbt_version: data.dbtVersion,
-                    semantic_layer_connection: encryptedSemanticLayerConnection,
+                    semantic_layer_connection: null,
                     ...(copiedProjects.length === 1
                         ? {
                               scheduler_timezone:
@@ -646,23 +628,6 @@ export class ProjectModel {
                     );
                 }
 
-                let semanticLayerConnection:
-                    | SemanticLayerConnection
-                    | undefined;
-                try {
-                    semanticLayerConnection = project.semantic_layer_connection
-                        ? (JSON.parse(
-                              this.encryptionUtil.decrypt(
-                                  project.semantic_layer_connection,
-                              ),
-                          ) as SemanticLayerConnection)
-                        : undefined;
-                } catch (e) {
-                    throw new UnexpectedServerError(
-                        'Failed to load dbt credentials',
-                    );
-                }
-
                 const result: Omit<Project, 'warehouseConnection'> = {
                     organizationUuid: project.organization_uuid,
                     projectUuid,
@@ -672,7 +637,6 @@ export class ProjectModel {
                     pinnedListUuid: project.pinned_list_uuid,
                     dbtVersion: project.dbt_version,
                     upstreamProjectUuid: project.copied_from_project_uuid,
-                    semanticLayerConnection,
                     schedulerTimezone: project.scheduler_timezone,
                     createdByUserUuid: project.created_by_user_uuid,
                 };
@@ -738,31 +702,6 @@ export class ProjectModel {
         };
     }
 
-    private static getSemanticLayerNonSensitiveCredentials(
-        semanticLayerConnection: SemanticLayerConnection,
-    ) {
-        const { type } = semanticLayerConnection;
-
-        switch (type) {
-            case SemanticLayerType.CUBE:
-                return {
-                    type,
-                    domain: semanticLayerConnection.domain,
-                } as CubeSemanticLayerConnection;
-            case SemanticLayerType.DBT:
-                return {
-                    type,
-                    domain: semanticLayerConnection.domain,
-                    environmentId: semanticLayerConnection.environmentId,
-                } as DbtSemanticLayerConnection;
-            default:
-                return assertUnreachable(
-                    type,
-                    `Unknown semantic layer connection type: ${type}`,
-                );
-        }
-    }
-
     /* 
     This method will load default values for backwards compatibility
     For example, when we introduce a new authentication type, we need to set the default value for the existing projects
@@ -812,8 +751,6 @@ export class ProjectModel {
     async get(projectUuid: string): Promise<Project> {
         const project = await this.getWithSensitiveFields(projectUuid);
         const sensitiveCredentials = project.warehouseConnection;
-        const sensitiveSemanticLayerCredentials =
-            project.semanticLayerConnection;
 
         const nonSensitiveDbtCredentials = Object.fromEntries(
             Object.entries(project.dbtConnection).filter(
@@ -832,12 +769,6 @@ export class ProjectModel {
                   ),
               ) as WarehouseCredentials)
             : undefined;
-        const nonSensitiveSemanticLayerCredentials =
-            sensitiveSemanticLayerCredentials
-                ? ProjectModel.getSemanticLayerNonSensitiveCredentials(
-                      sensitiveSemanticLayerCredentials,
-                  )
-                : undefined;
 
         const nonSensitiveCredentialsWithDefaults =
             ProjectModel.getConnectionWithDefaults(
@@ -855,7 +786,6 @@ export class ProjectModel {
             pinnedListUuid: project.pinnedListUuid,
             dbtVersion: project.dbtVersion,
             upstreamProjectUuid: project.upstreamProjectUuid || undefined,
-            semanticLayerConnection: nonSensitiveSemanticLayerCredentials,
             schedulerTimezone: project.schedulerTimezone,
             createdByUserUuid: project.createdByUserUuid ?? null,
         };
@@ -2388,77 +2318,6 @@ export class ProjectModel {
                     [name],
                 ),
             });
-    }
-
-    private static isSemanticLayerConnectionValid(
-        semanticLayerConnection: SemanticLayerConnectionUpdate,
-    ): boolean {
-        const { type } = semanticLayerConnection;
-        switch (type) {
-            case SemanticLayerType.DBT:
-                return Boolean(
-                    semanticLayerConnection.domain &&
-                        semanticLayerConnection.environmentId &&
-                        semanticLayerConnection.token,
-                );
-            case SemanticLayerType.CUBE:
-                return Boolean(
-                    semanticLayerConnection.domain &&
-                        semanticLayerConnection.token,
-                );
-            default:
-                return assertUnreachable(
-                    type,
-                    `Unknown semantic layer connection type: ${type}`,
-                );
-        }
-    }
-
-    async updateSemanticLayerConnection(
-        projectUuid: string,
-        connectionUpdate: SemanticLayerConnectionUpdate,
-    ) {
-        const { semanticLayerConnection: currentSemanticLayerConnection } =
-            await this.getWithSensitiveFields(projectUuid);
-
-        // ? Merging so the partial update only overwrites the existing properties, even when the current connection is undefined since merge will take ALL properties
-        const shouldMerge =
-            !currentSemanticLayerConnection ||
-            connectionUpdate.type === currentSemanticLayerConnection.type;
-
-        const updatedSemanticLayerConnection = shouldMerge
-            ? merge(currentSemanticLayerConnection, connectionUpdate)
-            : connectionUpdate;
-
-        if (
-            !ProjectModel.isSemanticLayerConnectionValid(
-                updatedSemanticLayerConnection,
-            )
-        ) {
-            throw new ParameterError('Invalid semantic layer connection');
-        }
-
-        const [updatedProject] = await this.database(ProjectTableName)
-            .update({
-                semantic_layer_connection: this.encryptionUtil.encrypt(
-                    JSON.stringify(updatedSemanticLayerConnection),
-                ),
-            })
-            .where('project_uuid', projectUuid)
-            .returning('*');
-
-        return updatedProject;
-    }
-
-    async deleteSemanticLayerConnection(projectUuid: string) {
-        const [updatedProject] = await this.database(ProjectTableName)
-            .update({
-                semantic_layer_connection: null,
-            })
-            .where('project_uuid', projectUuid)
-            .returning('*');
-
-        return updatedProject;
     }
 
     async updateDefaultSchedulerTimezone(