+++ packages/common/src/types/schedulerTaskList.ts
@@ -16,7 +16,6 @@ import {
     type TraceTaskBase,
     type ValidateProjectPayload,
 } from './scheduler';
-import { type SemanticLayerQueryPayload } from './semanticLayer';
 import {
     type SqlRunnerPayload,
     type SqlRunnerPivotQueryPayload,
@@ -38,7 +37,6 @@ export const SCHEDULER_TASKS = {
     COMPILE_PROJECT: 'compileProject',
     CREATE_PROJECT_WITH_COMPILE: 'createProjectWithCompile',
     TEST_AND_COMPILE_PROJECT: 'testAndCompileProject',
-    SEMANTIC_LAYER_QUERY: 'semanticLayer',
     SQL_RUNNER: 'sqlRunner',
     SQL_RUNNER_PIVOT_QUERY: 'sqlRunnerPivotQuery',
     REPLACE_CUSTOM_FIELDS: 'replaceCustomFields',
@@ -62,7 +60,6 @@ export interface TaskPayloadMap {
     [SCHEDULER_TASKS.COMPILE_PROJECT]: CompileProjectPayload;
     [SCHEDULER_TASKS.CREATE_PROJECT_WITH_COMPILE]: SchedulerCreateProjectWithCompilePayload;
     [SCHEDULER_TASKS.TEST_AND_COMPILE_PROJECT]: CompileProjectPayload;
-    [SCHEDULER_TASKS.SEMANTIC_LAYER_QUERY]: SemanticLayerQueryPayload;
     [SCHEDULER_TASKS.SQL_RUNNER]: SqlRunnerPayload;
     [SCHEDULER_TASKS.SQL_RUNNER_PIVOT_QUERY]: SqlRunnerPivotQueryPayload;
     [SCHEDULER_TASKS.REPLACE_CUSTOM_FIELDS]: ReplaceCustomFieldsPayload;